Refactor: Rename HR to Employee module
OPEN - task 88: Create HR module 
http://github.com/DevOpsDistilled/OpERP/issues/issue/88
