+++ OpERP/src/main/java/devopsdistilled/operp/server/context/RmiContext.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Import;
 import devopsdistilled.operp.server.context.account.AccountRmiContext;
 import devopsdistilled.operp.server.context.business.BusinessRmiContext;
 import devopsdistilled.operp.server.context.commons.CommonRmiContext;
-import devopsdistilled.operp.server.context.hr.HrRmiContext;
+import devopsdistilled.operp.server.context.employee.EmployeeRmiContext;
 import devopsdistilled.operp.server.context.items.ItemRmiContext;
 import devopsdistilled.operp.server.context.party.PartyRmiContext;
 import devopsdistilled.operp.server.context.stock.StockRmiContext;
@@ -14,7 +14,7 @@ import devopsdistilled.operp.server.context.stock.StockRmiContext;
 @Configuration
 @Import({ CommonRmiContext.class, ItemRmiContext.class, StockRmiContext.class,
 		PartyRmiContext.class, BusinessRmiContext.class,
-		AccountRmiContext.class, HrRmiContext.class })
+		AccountRmiContext.class, EmployeeRmiContext.class })
 public class RmiContext {
 
 }
