+++ OpERP/src/main/java/devopsdistilled/operp/server/data/repo/employee/EmployeeRepository.java
@@ -1,8 +1,8 @@
-package devopsdistilled.operp.server.data.repo.hr;
+package devopsdistilled.operp.server.data.repo.employee;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import devopsdistilled.operp.server.data.entity.hr.Employee;
+import devopsdistilled.operp.server.data.entity.employee.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
