+++ OpERP/src/main/java/devopsdistilled/operp/server/data/service/employee/EmployeeService.java
@@ -1,6 +1,6 @@
-package devopsdistilled.operp.server.data.service.hr;
+package devopsdistilled.operp.server.data.service.employee;
 
-import devopsdistilled.operp.server.data.entity.hr.Employee;
+import devopsdistilled.operp.server.data.entity.employee.Employee;
 import devopsdistilled.operp.server.data.service.EntityService;
 
 public interface EmployeeService extends EntityService<Employee, Long> {
